+++ torch_geometric/explain/explainer.py
@@ -59,7 +59,7 @@ class Explainer:
                 - :obj:`"attributes"`: Will mask each feature across all nodes.
 
         edge_mask_type (MaskType or str, optional): The type of mask to apply
-            on edges. Has the sample possible values as :obj:`node_mask_type`.
+            on edges. Has the same values as :obj:`node_mask_type`.
             (default: :obj:`None`)
         threshold_config (ThresholdConfig, optional): The threshold
             configuration.